requirements.txt
asgiref==3.8.1
bleach==6.3.0
Django==5.1.4
django-etc==1.4.0
django-hitcount==1.3.5
django-tinymce==5.0.0
pillow==12.2.0
python-dotenv==1.0.1
sqlparse==0.5.2
webencodings==0.5.1
whitenoise==6.8.2
